Split InternalDistributedJmxProxy into RMIJmxProxy and JolokiaJmxProxy - Issue #1605
Replace the monolithic InternalDistributedJmxProxy inner class with:
- AbstractDistributedJmxProxy: shared transport-agnostic logic
- RMIJmxProxy: native JMX MBean listener management
- JolokiaJmxProxy: Jolokia HTTP-based notification management
- RepairStatsParser: extracted repair stats parsing utility

Eliminates all isJolokiaEnabled conditionals, resolves PMD
CyclomaticComplexity and CPD duplication warnings.
